Move controller-runtime scheme registration to sample controller
