+++ controllers/suite_test.go
@@ -35,7 +35,6 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	componentv1alpha1 "github.com/kyma-project/template-operator/api/v1alpha1"
 	operatorkymaprojectiov1alpha1 "github.com/kyma-project/template-operator/api/v1alpha1"
 	//+kubebuilder:scaffold:imports
 )
@@ -87,10 +86,10 @@ var _ = BeforeSuite(func() {
 	Expect(err).NotTo(HaveOccurred())
 	Expect(cfg).NotTo(BeNil())
 
-	err = componentv1alpha1.AddToScheme(scheme.Scheme)
+	err = controllers.AddToScheme(scheme.Scheme)
 	Expect(err).NotTo(HaveOccurred())
 
-	err = operatorkymaprojectiov1alpha1.AddToScheme(scheme.Scheme)
+	err = controllers.AddToScheme(scheme.Scheme)
 	Expect(err).NotTo(HaveOccurred())
 
 	//+kubebuilder:scaffold:scheme
